# Tool overlay render optimisations