Closes #520

On my machine and my city saves about 12-25 ms (50-60ms jumps to 75-85 ms) per frame as long as the camera remains stationary.

- [x] Add two new utility classes for storing an array of something (node ids or segment ids) and for storing camera position and rotation
- [x] Use the new functionality in
  - [x] Lane Connector tool
  - [x] Speed Limits tool
  - [x] Parking Restrictions tool

Not used in Priority Signs tool and in Vehicle Restrictions tool because they have own different caching mechanism.