@aubergine10 2019-09-07
Tested:

- [x] Speed limits - drops 3fps while camera is static with lots of roads on screen (before was about 6fps)
- [x] Lane connector  - drops 1fps while camera is static (before was about 3fps)
- [ ] Parking restrictions - note to self: Don't update graphics drivers in the middle of fps testing