@aubergine10 2019-09-07
Retesting with new graphics drivers and saved camera positions (ensure same scene complexity)...

Before optimisations:

* Speed limits: 46 fps
* Lane connector: 29-37 fps
* Parking restrictions: 43 fps

After optimisations:

* Speed limits: 46 fps
* Lane connector: 46-47 fps
* Parking restrictions: 44 fps

So lane connector is certainly a lot faster.