@kvakvs 2019-09-07
The two others did already have some similar optimisation in place, so i just changed the code to use the same code in all 3. There should not be observed speed up in those others but good to know it is stable and so far is safe.