@krzychu124 2019-09-07
Conclusions:
- if I don't move camera it's fine. ✔️  
Almost no frame rate change when camera is not moving, but horrible lags when it does (sometimes drop from >30 to ~15-18)...