@aubergine10 2019-09-07
I think it would still be worth pushing this update as is, it makes lane connector much faster while camera is still and that's still useful for users. Ultimately we need that proper caching manager we discussed some time ago (see #415 and possibly duplicate #517).