@krzychu124 2019-09-07
I've made one more test because I was curious what is the exact difference in performance of `ShowOverlay()` between stationary camera and constantly moving with tool enabled(Speed limits).

My potato computer generated following result of benchmark:
_Stationary camera:_ __0.4-0.64 ms/call__
_Camera rotating around intersection:_ __5.1-6.6 ms/call__