@kvakvs 2019-09-08
After the yesterday's discussion i reverted my uncommitted changes.
What should be fixed here, or deliver it as is?