@aubergine10 2019-09-08
Everything in this PR is working fine, except the speed limits don't show on second usage if camera did not change from prior usage.