@kvakvs 2019-09-08
Fixed the speed limits by resetting the cached camera on tool activation.
This still will not show overlay until the camera moves but that is not a deal breaker and i hope that is something we can address separately when changing each tool.